Split ModalStepper in smaller files
